+++ dtool-lookup-webapp/src/components/TextSearch.vue
@@ -6,24 +6,54 @@
       v-model="textQuery"
       v-on:keyup.enter.prevent="startSearch"
     />
+    <span
+      v-if="textQuery !== ''"
+      :class="{ 'is-invalid': !isJson, 'is-valid': isJson }"
+      >{{ isJson ? "✔ JSON Format" : "✘ Not in JSON" }}</span
+    >
   </form>
 </template>
 
 <script>
 export default {
   name: "TextSearch",
-  data: function() {
+  data: function () {
     return {
-      textQuery: ""
+      textQuery: "",
+      isFocused: false,
     };
   },
+  computed: {
+    isJson: function () {
+      if (this.textQuery === "") {
+        return false;
+      }
+      try {
+        JSON.parse(this.textQuery);
+        return true;
+      } catch (e) {
+        return false;
+      }
+    },
+  },
   methods: {
-    startSearch: function() {
-      this.$store.commit("update_free_text", this.textQuery);
+    startSearch: function () {
+      if (this.isJson) {
+        this.$store.commit("update_mongo_text", this.textQuery);
+      } else {
+        this.$store.commit("update_free_text", this.textQuery);
+      }
       this.$emit("start-search");
-    }
-  }
+    },
+  },
 };
 </script>
 
-<style></style>
+<style>
+.is-invalid {
+  color: red;
+}
+.is-valid {
+  color: rgb(0, 255, 0);
+}
+</style>