+++ dtool-lookup-webapp/src/main.js
@@ -12,6 +12,7 @@ import "bootstrap-vue-3/dist/bootstrap-vue-3.css";
 
 const app = createApp(App);
 
+
 app.use(BootstrapVue3);
 app.use(VueClipboard);
 app.use(store);